CSELLS-25: Create Hashes component
